rename copy tests
