+++ test/h5cpp_tests/node_tests/__init__.py
@@ -4,7 +4,7 @@ from .dataset_direct_chunk_test import *
 from .group_creation_test import *
 from .dataset_extent_test import *
 from .h5py_compat_test import *
-from .copy_tests import *
+from .copy_test import *
 from .link_iteration_test import *
 from .node_iterator_test import *
 from .virtual_dataset_test import *
